rewrite using context
