+++ src/components/Footer/Footer.tsx
@@ -74,7 +74,7 @@ export const Footer: React.FC<Props> = ({
         className="todoapp__clear-completed"
         data-cy="ClearCompletedButton"
         onClick={onClearCompleted}
-        disabled={completedTodos.length === 0}
+        disabled={!completedTodos.length}
       >
         Clear completed
       </button>
